pyproject.toml
[project]
name = "anime-presenter"
version = "0.1.0"
description = ""
authors = [{ name = "VLadislao", email = "VladX09rus@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyvidplayer2 (>=0.9.25,<0.10.0)",
    "typer (>=0.15.2,<0.16.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
packages = [{ include = "anime_presenter", from = "src" }]

[tool.poetry.scripts]
anime-presenter = "anime_presenter.cli:app"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
flake8 = "^7.1.2"
mypy = "^1.15.0"
pytest = "^8.3.5"
isort = "^6.0.1"
black = { extras = ["jupyter"], version = "^25.1.0" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[tool.mypy]
ignore_missing_imports = true

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'
